// src/routes/screenName.enum.ts
enum ScreenNameEnum {
  //Auth Screens
  SPLASH_SCREEN = 'SPLASH_SCREEN',
  OnboardingScreen = "OnboardingScreen",
  PhoneLogin = "PhoneLogin",
  SocialLogin = "SocialLogin",
  HomeStack = "HomeStack",
  ProfileScreen = "ProfileScreen",
  NotificationsScreen = "NotificationsScreen",
  UploadDocumentsScreen = "UploadDocumentsScreen",
  DeliveryHome = "DeliveryHome",
  PickupLocationRapido = "PickupLocationRapido",
  AllOrder = "AllOrder",
  NearbyDriversMap = "NearbyDriversMap",
  RequestLoading = "RequestLoading",
  CreateParcelFrom = "CreateParcelFrom",
  DeliveryRequest = "DeliveryRequest",
  EditProfileDeliver = "EditProfileDeliver",
  ChooseRole = "ChooseRole",
  AddressScreen = "AddressScreen",
  DeliveryTabNavigator = "DeliveryTabNavigator",
  EditProfile = "EditProfile",
  VehicleSetupScreen = "VehicleSetupScreen",
  BankSetupScreen = "BankSetupScreen",
  PickupLocation = "PickupLocation",
  ParcelDetails = "ParcelDetails",
  HelpSupport = "HelpSupport",
  changePassword = "ChangePassword",
  WalletScreen = "WalletScreen",
  CourierTrackingScreen = "CourierTrackingScreen",
  OrdersPrfile = "OrdersPrfile",
  ChatScreen = "ChatScreen",
  ContactInfo = "ContactInfo",
  PrivacyPolicy = "PrivacyPolicy",
  LegalPoliciesScreen = "LegalPoliciesScreen",
  ViewDetails = "ViewDetails",
  DashBoardTwo = "DashBoardTwo",
  NotificationsSetting = "NotificationsSetting",
  Sinup = "Sinup",
  ProfileSetup = "ProfileSetup",
  PickupFromLocation = "PickupFromLocation",
  Help = "Help",
  OfferOR = "OfferOR",
  Paymentinformation = "Paymentinformation",
  HelpDetail = "HelpDetail",
  DocumentShow = "DocumentShow",
  Step3Screen = "Step3Screen",
  UserInfo = "UserInfo",
  PrivacyPocily = "PrivacyPocily",
  PaymentHistoryScreen = "PaymentHistoryScreen",
  TabNavigator = "TabNavigator",
  DashBoardScreen = "DashBoardScreen",
  EarningsScreen = "EarningsScreen",
  ReadyScreen = "ReadyScreen",
  LoginScreen = "LoginScreen",
  TimeClock = "TimeClock",
  GeneralInfo = "GeneralInfo",
  OtpScreen = "OtpScreen",
  CreatePassword = "CreatePassword",
  PasswordReset = "PasswordReset",
  DashBoardDetail = "DashBoardDetail",
  PatientScreen = "PatientScreen",
  DrawerNavgation = "DrawerNavgation",
  NewTripScreen = 'NewTripScreen',
  AllDriverMapScreen = 'AllDriverMapScreen',
  AddPatient = "AddPatient",
  AddContract = "AddContract",
  AddDriver = "AddDriver",
  AddCar = "AddCar",
  AddPresubscription = "AddPresubscription",
  SuccessScreen = "SuccessScreen",
  MapScreen = "MapScreen",
  Billing = "Billing",
  patientDetail = "PatientDetail",
  driverDetail = "driverDetail",
  contractDetail = "contractDetail",
  setting = "Setting",
  language = "Language",
  personalInfo = "PersonalInfo",
  card = "card",
  condition = "Conditions",
  socity = "Socity",
  AddContractWithType = "AddContractWithType",
  WebViewScreen = "WebViewScreen",
  // Driver

  DriverHome = "DriverHome",
  TripDetail = "TripDetail",
  TripSuccess = "TripSuccess",
  TripMap = "TripMap",
  CaptureDoc = "CaptureDoc",
  CaptureSuccess = "CaptureSuccess",
  Patient_Driver = "Patient_Driver",
  RaceDetail = "RaceDetail",
  PresubscriberDetail = "presubscriberDetail",
  CarDetail = "CarDetail",
  VerificationPending = "VerificationPending"
}

export default ScreenNameEnum;